Monkeypatch the cucumber stack filters
This function is used to calculate failure errors paths to steps.
Without this patch, the decorators over step functions are displayed,
instead of the code implementing the step.

This change involves touching internal state of cucumber, and that is a
tradeoff with user experience.

For users of the library, should be easier to debug broken tests,
instead of purity on the codebase.