Fix problem by implicit @types/node poluting types
For pure frontend projects, implicitly using node types generates
clashes with functions like `setInterval` and `setTimeout` because the
interfaces are different.

This PR uses a more modern typescript compiler and also uses
`"types": []` on `tsconfig.json` to prevent implicit types from being used.

This was tested manually in a pure frontend project and works fine,
without this patches an error was triggered.